+++ libs/blocks/src/lib/navigation/navigation.link.tsx
@@ -1,6 +1,7 @@
 import { CustomLink } from '@deriv-com/components';
 import { DirectNavLink } from '@deriv-com/providers';
 import { Text, qtJoin, qtMerge } from '@deriv/quill-design';
+import { LabelPairedArrowUpRightRegularIcon } from '@deriv/quill-icons/LabelPaired';
 
 export interface NavigationLinkProps {
   item: DirectNavLink;
@@ -15,17 +16,24 @@ export const NavigationLink: React.FC<NavigationLinkProps> = ({
     <CustomLink
       href={item.href}
       className={qtMerge(
-        'justify-start gap-gap-md text-body-md font-bold leading-body-md text-typography-default',
+        'w-full justify-start gap-gap-md py-general-sm font-bold',
         item.icon && 'grid grid-cols-[24px_1fr]',
         !item.icon && item.description && 'flex flex-col',
         className,
       )}
     >
       {item.icon ? item.icon : null}
-      {item.text}
-      <Text size="sm" className={qtJoin(item.icon && 'col-start-2')}>
-        {item.description}
-      </Text>
+      <div className="flex gap-gap-sm">
+        <Text size="md" bold>
+          {item.text}
+        </Text>
+        {item.external && <LabelPairedArrowUpRightRegularIcon iconSize="sm" />}
+      </div>
+      {item.description ? (
+        <Text size="sm" className={qtJoin(item.icon && 'col-start-2')}>
+          {item.description}
+        </Text>
+      ) : null}
     </CustomLink>
   );
 };