+++ libs/blocks/src/lib/social-proof/index.tsx
@@ -1,0 +1,21 @@
+import { ReactNode } from 'react';
+import SocialProofVertical from './vertical';
+import SocialProofHorizontal from './horizontal';
+
+type SocialProofVariants = {
+  Vertical: typeof SocialProofVertical;
+  Horizontal: typeof SocialProofHorizontal;
+};
+
+export const SocialProof: SocialProofVariants = ({
+  children,
+}: {
+  children: ReactNode;
+}) => {
+  return children;
+};
+
+SocialProof.Vertical = SocialProofVertical;
+SocialProof.Horizontal = SocialProofHorizontal;
+
+export default SocialProof;