+++ libs/components/src/index.ts
@@ -9,3 +9,4 @@ export * from './lib/link';
 export * from './lib/table/account-table';
 export * from './lib/action-sheet/bottom-sheet';
 export * from './lib/table-cell';
+export * from './lib/rating';