+++ libs/hooks/src/index.ts
@@ -1,3 +1,4 @@
 export * from './lib/use-build-variant';
 export * from './lib/use-navigation';
 export * from './lib/use-shared-link';
+export * from './lib/use-trustpilot';